fix: lazy loading & caching of s3 client
